Mention key restrictions
A natural instinct would be to follow the guidelines in the GCP console and to place an Android app restriction on the key. This will not work as the regular HTTP interface is used, where the app is not identified as such. HTTP referrer restrictions also do not work, because none are passed to the API.

Ultimately, if any application restrictions are placed on the key, requests will fail.